feat: 리스트 이미지 클릭 시 틀 고정 완료
리스트 이미지 클릭 시 기본 틀이 나오게 제작하였습니다.
